Move SierraAccessStatus into the transformer; it's not used elsewhere
