+++ pipeline/transformer/transformer_sierra/src/main/scala/weco/pipeline/transformer/sierra/transformers/SierraItems.scala
@@ -13,7 +13,6 @@ import weco.catalogue.internal_model.locations.{
 }
 import weco.catalogue.internal_model.work.Item
 import weco.catalogue.source_model.sierra.rules.{
-  SierraAccessStatus,
   SierraItemAccess,
   SierraPhysicalLocationType
 }
