+++ pipeline/transformer/transformer_sierra/src/main/scala/weco/pipeline/transformer/sierra/transformers/SierraLocation.scala
@@ -2,7 +2,6 @@ package weco.pipeline.transformer.sierra.transformers
 
 import weco.catalogue.internal_model.locations._
 import weco.catalogue.source_model.sierra.rules.{
-  SierraAccessStatus,
   SierraItemAccess,
   SierraPhysicalLocationType
 }
